requirements.txt
aiohttp
aiomysql
aiosqlite
discord.py
dotenv
emoji
flask
numpy
pillow
pytz
requests
topggpy
